Updates
- New file structure
- Removing old trinamic files
- moving disable motor code
