+++ Grbl_Esp32/Machines/motor_class_test.h
@@ -103,8 +103,8 @@
 
 #define ENABLE_STALLGUARD_TUNING
 
-//#define TRINAMIC_RUN_MODE           TRINAMIC_RUN_MODE_STEALTHCHOP
-#define TRINAMIC_RUN_MODE           TRINAMIC_RUN_MODE_COOLSTEP
+#define TRINAMIC_RUN_MODE           TRINAMIC_RUN_MODE_STEALTHCHOP
+//#define TRINAMIC_RUN_MODE           TRINAMIC_RUN_MODE_COOLSTEP
 //#define TRINAMIC_RUN_MODE           TRINAMIC_RUN_MODE_STALLGUARD
 
 #define TRINAMIC_HOMING_MODE        TRINAMIC_HOMING_STALLGUARD
